chore(coupon-kafka-consumer): move appropriate packages
